GPT/requirements.txt
streamlit
pandas
openai
imaplib2
python-dotenv
smtplib
email